Remove activity table
Rethinking this all a bit. Don't want to rely on a database for actual user interactions, it doesn't scale and I don't want to maintain it. So these database are only going to be used for processing and dumping the data so it can be used elsewhere. User interactions will have to happen some other clever way.